+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.1.9 - 2024-02-22
+
 - roll back compile API to 33 to keep back compatibility ([#98](https://github.com/PostHog/posthog-android/pull/98))
 - set `device_type` accordingly to Mobile, TV, or Tablet ([#93](https://github.com/PostHog/posthog-android/pull/93))